# developer-docs/openapi.yaml
openapi: 3.1.0
info:
  title: ShipCheck API
  version: 0.1.0
  description: Requirement-to-proof acceptance service for public web deliveries.
servers:
  - url: https://api.shipcheck.example
paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /v1/verify:
    post:
      summary: Compile a brief, execute supported checks, and return an acceptance receipt
      operationId: verifyDelivery
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./schemas/verify-request.schema.json
      responses:
        "200":
          description: Verification completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyResponse"
        "400":
          description: Invalid request or unsafe target
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "402":
          description: x402 payment required
        "409":
          description: Idempotency key conflict or request already in progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Requirement compilation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Safe execution unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/compile:
    post:
      summary: Compile a brief into an acceptance contract without executing it
      operationId: compileRequirements
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./schemas/verify-request.schema.json
      responses:
        "200":
          description: Requirements compiled
          content:
            application/json:
              schema:
                $ref: ./schemas/acceptance-contract.schema.json
        "400":
          description: Invalid request or unsafe target
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Requirement compilation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/requests/{requestId}:
    get:
      summary: Retrieve verification request status
      operationId: getVerificationRequest
      parameters:
        - in: path
          name: requestId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Verification request status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestStatusResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/receipts/{receiptId}:
    get:
      summary: Retrieve an acceptance receipt
      parameters:
        - in: path
          name: receiptId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Acceptance receipt
          content:
            application/json:
              schema:
                $ref: ./schemas/acceptance-receipt.schema.json
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/receipts/{receiptId}/verify:
    get:
      summary: Verify receipt hashes and optional anchor
      parameters:
        - in: path
          name: receiptId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceiptVerificationResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/reports/{receiptId}:
    get:
      summary: Fetch an unlisted human-readable report bundle
      description: Reports are accessed by unguessable receipt ID. Brief text and raw page content are not included.
      parameters:
        - in: path
          name: receiptId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Report bundle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportBundleResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/reports/{receiptId}/evidence/{evidenceId}/link:
    get:
      summary: Create a short-lived evidence read link
      parameters:
        - in: path
          name: receiptId
          required: true
          schema:
            type: string
        - in: path
          name: evidenceId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Expiring evidence link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenceLinkResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthResponse:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          const: ok
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
    ErrorDetail:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - INVALID_REQUEST
            - UNSAFE_TARGET
            - COMPILATION_FAILED
            - REQUEST_CONFLICT
            - NOT_FOUND
            - RATE_LIMITED
            - EXECUTION_UNAVAILABLE
            - INTERNAL_ERROR
        message:
          type: string
        details: {}
        requestId:
          type: string
    VerifySummary:
      type: object
      additionalProperties: false
      required: [total, passed, failed, unverified, notObjectivelyTestable, unsupported]
      properties:
        total:
          type: integer
          minimum: 0
        passed:
          type: integer
          minimum: 0
        failed:
          type: integer
          minimum: 0
        unverified:
          type: integer
          minimum: 0
        notObjectivelyTestable:
          type: integer
          minimum: 0
        unsupported:
          type: integer
          minimum: 0
    VerifyResponse:
      type: object
      additionalProperties: false
      required: [requestId, contract, verdict, summary, results, receipt]
      properties:
        requestId:
          type: string
        contract:
          $ref: ./schemas/acceptance-contract.schema.json
        verdict:
          type: string
          enum: [ACCEPTED, ACCEPTED_WITH_NOTES, CHANGES_REQUIRED, INSUFFICIENT_SPECIFICATION, EXECUTION_INCOMPLETE]
        summary:
          $ref: "#/components/schemas/VerifySummary"
        results:
          type: array
          items:
            type: object
        receipt:
          $ref: ./schemas/acceptance-receipt.schema.json
        reportUrl:
          type: string
          format: uri
    RequestStatusResponse:
      type: object
      additionalProperties: false
      required: [requestId, status]
      properties:
        requestId:
          type: string
        status:
          type: string
          enum: [QUEUED, RUNNING, COMPLETED, FAILED]
        response:
          $ref: "#/components/schemas/VerifyResponse"
        error:
          $ref: "#/components/schemas/ErrorDetail"
    ReceiptVerificationResponse:
      type: object
      additionalProperties: false
      required: [receiptId, valid, checks]
      properties:
        receiptId:
          type: string
        valid:
          type: boolean
        checks:
          type: object
          additionalProperties:
            type: boolean
    ReportBundleResponse:
      type: object
      additionalProperties: false
      required: [receiptId, contract, verdict, summary, results, receipt, evidenceManifest, createdAt]
      properties:
        receiptId:
          type: string
        contract:
          $ref: ./schemas/acceptance-contract.schema.json
        verdict:
          type: string
        summary:
          $ref: "#/components/schemas/VerifySummary"
        results:
          type: array
          items:
            type: object
        receipt:
          $ref: ./schemas/acceptance-receipt.schema.json
        evidenceManifest:
          $ref: ./schemas/evidence-manifest.schema.json
        createdAt:
          type: string
          format: date-time
    EvidenceLinkResponse:
      type: object
      additionalProperties: false
      required: [receiptId, evidenceId, url, expiresAt]
      properties:
        receiptId:
          type: string
        evidenceId:
          type: string
        url:
          type: string
          format: uri
        expiresAt:
          type: string
          format: date-time